==> 01_Building_Dataset_v5.ipynb <==
# Création du dataset supervisé

- Ce notebook a pour objectif de créer un dataset supervisé pour entraîner l'algorithme de scoring
- Les clients qui n'ont jamais achetés de produits sont donc supprimés (cold start customers) et peuvent être traités séparrément (ex: recommandation des produits les plus populaires)
- Les données sont converties au format numeric car l'opérateur QueryFeast de Triton ne supporte pas les String pour le moment

import os
import gc
import glob
os.environ["TF_GPU_ALLOCATOR"]="cuda_malloc_async"

import nvtabular as nvt
from nvtabular.ops import *
import numpy as np

from merlin.models.utils.example_utils import workflow_fit_transform
from merlin.schema.tags import Tags
from merlin.models.utils.dataset import unique_rows_by_features

import merlin.models.tf as mm
from merlin.io.dataset import Dataset
import tensorflow as tf

# Disable INFO and DEBUG logging everywhere
import logging

logging.disable(logging.WARNING)

INPUT_DATA_DIR = os.environ.get("INPUT_DATA_DIR", "/root/Data/Row/")
DATA_FOLDER = os.environ.get("DATA_FOLDER", "/root/Data/")
MODELS_FOLDER = os.environ.get("MODELS", "/root/Models/")
PROCESSED_FOLDER = os.environ.get("PROCESSED_FOLDER", "/root/Data/Processed/")
feature_repo_path = os.environ.get("FEAST_PATH", "/root/Data/feast_repo/feature_repo")

BATCH_SIZE = int(os.environ.get("BATCH_SIZE", 512))
from merlin.core.dispatch import get_lib
df_lib = get_lib()
df_lib

## Chargement des datasets

# Modifie les dtypes qui peuvent être en 64 par défaut

def change_dtypes(df):
    for col in df.columns:
        if df[col].dtype == np.int64:
            df[col] = df[col].astype(np.int32)
        elif df[col].dtype == np.float64:
            df[col] = df[col].astype(np.float32)
    return df

transactions = df_lib.read_csv(os.path.join(DATA_FOLDER, "Row/transactions_train.csv"))

# Pour rendre le démonstrateur plus léger, on ne garde que les 25% du dataset les plus récents 
split = int(len(transactions)/4)*0
transactions = transactions[split:]
transactions['t_dat'] = df_lib.to_datetime(transactions['t_dat'])


customers = df_lib.read_csv(os.path.join(DATA_FOLDER, "Row/customers.csv"))
articles = df_lib.read_csv(os.path.join(DATA_FOLDER, "Row/articles.csv"))

transactions = change_dtypes(transactions)
customers = change_dtypes(customers)
articles = change_dtypes(articles)

Remplissage des valeurs manquantes

customers['FN'] = customers['FN'].fillna(2)
customers['Active'] = customers['Active'].fillna(2)
customers['club_member_status'] = customers['club_member_status'].fillna('NA')
customers['fashion_news_frequency'] = customers['fashion_news_frequency'].fillna('NA')
customers['age'] = customers['age'].fillna(35) #the mean
articles['detail_desc'] = articles['detail_desc'].fillna('NA')
transactions['article_id'] = transactions['article_id'].fillna(0)
transactions['price'] = transactions['price'].fillna(transactions['price'].mean())
transactions['sales_channel_id'] = transactions['sales_channel_id'].fillna(3)

Suppression des colonnes redondantes

articles = articles.drop(['product_type_name', 'graphical_appearance_name', 'colour_group_name', 'perceived_colour_value_name', 'department_name', 'index_name', 'index_group_name', 'section_name', 'garment_group_name', 'perceived_colour_master_name'], axis=1)

print('transactions :', transactions.columns)
print('customers :', customers.columns)
print('articles :', articles.columns)

## Construction de features supplémentaires

Pour les items, on ajoute des features pour capturer la popularité et les tendances récentes comme la date depuis la dernière transaction etc.

## Seule ligne à run 
#Time in days since transactions
Last_day = transactions['t_dat'].max()
transactions['days_diff'] = (Last_day - transactions['t_dat']).dt.days

#In the last 30 days before test period
count_30d_purchased = transactions[(transactions['days_diff'] <= 50) & (transactions['days_diff'] >=20)]['article_id'].value_counts().reset_index()
count_30d_purchased.columns = ['article_id', 'count_30d_purchased']

#In the last 7 days before test period
count_7d_purchased = transactions[(transactions['days_diff'] <= 27) & (transactions['days_diff'] >= 20)]['article_id'].value_counts().reset_index()
count_7d_purchased.columns = ['article_id', 'count_7d_purchased']

#Time weighted purchased
filtered_transactions = transactions[transactions['days_diff'] >= 20]

daily_transactions = filtered_transactions.groupby(['article_id', 'days_diff']).size().reset_index()
daily_transactions.columns = ['article_id', 'days_diff', 'nbr_sales']
daily_transactions['Time_Weighted_sales'] = daily_transactions['nbr_sales']/(daily_transactions['days_diff']+1)
Time_Weighted_Purchased = daily_transactions.groupby('article_id')['Time_Weighted_sales'].sum().reset_index()
Time_Weighted_Purchased.columns = ['article_id', 'Time_Weighted_Purchased']

#Merge
articles = articles.merge(count_30d_purchased, on='article_id', how='left')
articles = articles.merge(count_7d_purchased, on='article_id', how='left')
articles = articles.merge(Time_Weighted_Purchased, on='article_id', how='left')

#Fill missing values
articles['count_30d_purchased'] = articles['count_30d_purchased'].fillna(0)
articles['count_7d_purchased'] = articles['count_7d_purchased'].fillna(0)
articles['Time_Weighted_Purchased'] = articles['Time_Weighted_Purchased'].fillna(0)

articles.head(1)

Pour les clients, on ajoute dans un premier temps les features classiques de segmentation marketing : Recency, Frequency et Amount

#Recency
filtered_transactions = transactions[transactions['days_diff'] >= 20]
Recency = filtered_transactions.groupby('customer_id')['days_diff'].min().reset_index()
Recency.columns = ['customer_id', 'recency']

#Frequency
Frequency = transactions[(transactions['days_diff'] <=100) & (transactions['days_diff'] >= 20)].groupby('customer_id').size().reset_index()
Frequency.columns = ['customer_id', 'frequency']
Frequency['frequency'] = Frequency['frequency']/100

#Amount
Amount = transactions[(transactions['days_diff'] <= 100) & (transactions['days_diff'] >= 20)].groupby('customer_id')['price'].mean().reset_index()
Amount.columns = ['customer_id', 'amount']

#Merge
customers = customers.merge(Recency, on='customer_id', how='left')
customers = customers.merge(Frequency, on='customer_id', how='left')
customers = customers.merge(Amount, on='customer_id', how='left')

#Fill missing values
#if some customers have not ordered, we set their recency to more than one year
customers['recency'] = customers['recency'].fillna(500)
customers['frequency'] = customers['frequency'].fillna(0)
customers['amount'] = customers['amount'].fillna(0)

customers.head(2)

On ajoute ensuite des features qui capturent les goûts et préférences des clients comme la catégorie la plus fréquente des articles achetés etc.

Pour ce faire on liste tous les derniers achats par clients.

transactions_legit =  transactions[transactions['days_diff'] >= 20]
merged = transactions_legit[['t_dat', 'customer_id', 'article_id']].merge(articles[['article_id', 'product_code', 'product_type_no', 'colour_group_code', 'department_no', 'section_no']], on='article_id')
merged = merged.reset_index(drop=True)
#merged.head(2)

grouped = merged.groupby('customer_id').agg({'article_id': 'collect', 'product_code': 'collect', 'product_type_no': 'collect', 'colour_group_code': 'collect', 'department_no': 'collect',
                                                       'section_no': 'collect'})
grouped.head(2)

Quelques fonctions pour extraire les éléments les plus fréquents ou les plus récents dans les listes d'items achetés par clients

from collections import Counter

def most_frequent(lst):
    if len(lst) > 0:
        return Counter(lst).most_common(1)[0][0]
    return 0

def second_most_frequent(lst):
    most_common = Counter(lst).most_common(2)
    if len(most_common) > 1:
        return most_common[1][0]
    return 0

def last_n_elements(lst, n):
    if len(lst) >= n:
        return lst[-n]
    return 0

grouped = grouped.to_pandas()

%%time
user_additional_features_df = df_lib.DataFrame({
    'customer_id': grouped.index,
    'popular_product_type': grouped['product_type_no'].apply(most_frequent),
    '2nd_popular_product_type': grouped['product_type_no'].apply(second_most_frequent),
    'popular_department_no': grouped['department_no'].apply(most_frequent),
    '2nd_popular_department_no': grouped['department_no'].apply(second_most_frequent),
    'popular_section_no': grouped['section_no'].apply(most_frequent),
    '2nd_popular_section_no': grouped['section_no'].apply(second_most_frequent),
    'last_product_code': grouped['product_code'].apply(lambda x: last_n_elements(x, 1)),
    '2nd_last_product_code': grouped['product_code'].apply(lambda x: last_n_elements(x, 2)),
    #'last_article_id': grouped['article_id'].apply(lambda x: last_n_elements(x, 1)),
    #'2nd_last_article_id': grouped['article_id'].apply(lambda x: last_n_elements(x, 2)),
    'last_product_type': grouped['product_type_no'].apply(lambda x: last_n_elements(x, 1)),
    '2nd_last_product_type': grouped['product_type_no'].apply(lambda x: last_n_elements(x, 2)),
})
user_additional_features_df.head(2)

customers = customers.merge(user_additional_features_df, on='customer_id')
customers

print('transactions :', transactions.columns)
print('customers :', customers.columns)
print('articles :', articles.columns)

articles.isna().sum()

## Création du dataset
Création de faux négatifs

transactions = transactions.rename(columns={'article_id': 'item_id', 'customer_id': 'user_id'})
customers = customers.rename(columns={'customer_id': 'user_id'})
articles = articles.rename(columns={'article_id': 'item_id'})

true_set = transactions[transactions['days_diff'] < 20][['user_id', 'item_id']].copy()
true_set['Target'] = 1

all_articles = transactions['item_id'].unique().to_arrow().to_pylist()
all_customers = transactions['user_id'].unique().to_arrow().to_pylist()

# Calculer les poids pour les articles et les utilisateurs
often_purchased = articles[articles['count_30d_purchased'] > 15]
article_weights = often_purchased[often_purchased['item_id'].isin(all_articles)].groupby('item_id')['count_30d_purchased'].max()

# Inverser les poids pour les utilisateurs pour que les recency faibles aient des poids élevés
often_customers = customers[customers['recency'] > 0]
user_weights = often_customers[often_customers['user_id'].isin(all_customers)].groupby('user_id')['recency'].max() + 1.5
user_weights = 1 / (user_weights + 1)

# Normaliser les poids pour obtenir des probabilités
article_weights = article_weights / article_weights.sum()
user_weights = user_weights / user_weights.sum()

# Nouvelle liste de users et d'items
good_articles = often_purchased[often_purchased['item_id'].isin(all_articles)]['item_id'].unique().to_arrow().to_pylist()
good_customers = often_customers[often_customers['user_id'].isin(all_customers)]['user_id'].unique().to_arrow().to_pylist()


false_set = df_lib.DataFrame(columns=['user_id', 'item_id', 'Target'])
batch_size = 10000  # adjust this based on your memory capacity

# Create a unique id for each pair in true_set
true_set['pair_id'] = true_set['item_id'].astype(str).str.cat(true_set['user_id'].astype(str), sep='_')

while len(false_set) < len(true_set): # positive rate
    # Select a batch of random rows from true_set
    random_rows = true_set.sample(n=batch_size).copy()
   
    # Select a batch of random users and items
    random_users = np.random.choice(good_customers, size=batch_size, replace=True, p=user_weights.loc[good_customers].values.get())
    random_items = np.random.choice(good_articles, size=batch_size, replace=True, p=article_weights.loc[good_articles].values.get())
    
    # Create a new DataFrame for the false negatives
    random_rows['user_id'] = random_users
    random_rows['item_id'] = random_items
    random_rows['Target'] = 0

    # Create a unique id for each pair in false_negatives
    random_rows['pair_id'] = random_rows['item_id'].astype(str).str.cat(random_rows['user_id'].astype(str), sep='_')

    # Check that the pairs are not in true_set
    random_rows = random_rows[~random_rows['pair_id'].isin(true_set['pair_id'])]

    false_set = df_lib.concat([false_set, random_rows])

# Concatenate true positives and false negatives to get the final dataset
dataset = df_lib.concat([true_set, false_set])
#Shuffle the dataset
dataset = dataset.sample(frac=1)

# Drop the pair_id column
dataset = dataset.drop(columns='pair_id')

dataset['Target'] = dataset['Target'].astype('int32')

import gc
del true_set
del false_set
gc.collect()

dataset

dataset = dataset.merge(articles, on='item_id', how='left')
#print(dataset.isna().sum())

print(dataset['user_id'].nunique())

valid_user_ids = customers['user_id'].unique()
filtered_dataset = dataset[dataset['user_id'].isin(valid_user_ids)]
print(filtered_dataset['user_id'].nunique())

merged_dataset = customers.merge(filtered_dataset, on='user_id', how='right')

dataset = merged_dataset
merged_dataset.isna().sum()

dataset.info()

dataset.head(2)

dataset.columns

## Conversion en train-test numeric

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder


# Séparer les caractéristiques et la cible
dataset['user_row'] = dataset['user_id'].copy()
dataset = change_dtypes(dataset)
X = dataset.drop('Target', axis=1)
y = dataset['Target']

# Diviser les données en ensembles d'entraînement et de test
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)

X_train.columns

categorical_columns = ['prod_name', 'product_group_name', 'index_code', 'detail_desc', 'FN', 'Active', 'club_member_status', 'fashion_news_frequency', 'postal_code']
other_features = ['user_id', 'item_id', 'product_code', 'product_type_no', 'graphical_appearance_no', 'colour_group_code',
       'perceived_colour_value_id', 'perceived_colour_master_id',
       'department_no', 'index_group_no', 'section_no',
       'garment_group_no', 'count_30d_purchased',
       'count_7d_purchased', 'Time_Weighted_Purchased', 'age',
       'recency', 'frequency', 'amount', 'user_row',  'popular_product_type',
       '2nd_popular_product_type', 'popular_department_no',
       '2nd_popular_department_no', 'popular_section_no',
       '2nd_popular_section_no', 'last_product_code', '2nd_last_product_code',
       'last_product_type', '2nd_last_product_type']
print(len(categorical_columns)+len(other_features))
print(len(X_train.columns))

# categorify
features_cat = categorical_columns >> Categorify(dtype="int32") 
features = features_cat + other_features #>> Dropna()

workflow_encode = nvt.Workflow(features)

train_dataset = nvt.Dataset(X_train)
test_dataset = nvt.Dataset(X_test)

# transform
workflow_encode.fit(train_dataset)

train_transformed = workflow_encode.transform(train_dataset).to_ddf().compute()
test_transformed = workflow_encode.transform(test_dataset).to_ddf().compute()

print(len(train_transformed), len(y_train))

train_transformed.isna().sum()

train_transformed.info()

train_transformed

### Sauvegarde des datasets pour test sklearn

#dataset.to_parquet(DATA_FOLDER+'/Processed/dataset_numeric')
train_transformed.to_csv(DATA_FOLDER+'/Plot/train_numeric')
test_transformed.to_csv(DATA_FOLDER+'/Plot/test_numeric')
y_train.to_frame(name='Target').to_csv(DATA_FOLDER+'/Plot/ytrain_numeric')
y_test.to_frame(name='Target').to_csv(DATA_FOLDER+'/Plot/ytest_numeric')

y_train

### Sauvegarde des datasets Merlin

train_transformed = train_transformed.reset_index(drop=True)
y_train = y_train.reset_index(drop=True)
train_to_merlin = df_lib.concat([train_transformed, y_train], axis=1)

test_transformed = test_transformed.reset_index(drop=True)
y_test = y_test.reset_index(drop=True)
test_to_merlin = df_lib.concat([test_transformed, y_test], axis=1)

print(train_to_merlin.columns)

train_to_merlin.head(3)

customers_features = ['FN', 'Active', 'club_member_status',
       'fashion_news_frequency', 'age', 'postal_code', 'recency', 'frequency',
       'amount', 'popular_product_type', '2nd_popular_product_type',
       'popular_department_no', '2nd_popular_department_no',
       'popular_section_no', '2nd_popular_section_no', 'last_product_code',
       '2nd_last_product_code', 'last_product_type', '2nd_last_product_type']

articles_features = ['product_code', 'prod_name', 'product_type_no',
       'product_group_name', 'graphical_appearance_no', 'colour_group_code',
       'perceived_colour_value_id', 'perceived_colour_master_id',
       'department_no', 'index_code', 'index_group_no', 'section_no',
       'garment_group_no', 'detail_desc', 'count_30d_purchased',
       'count_7d_purchased', 'Time_Weighted_Purchased']

print(len(customers_features)+ len(articles_features), len(train_transformed.columns))

tag_user_id = ["user_id"] >> Categorify(dtype='int32') >> TagAsUserID()
tag_user_features = customers_features  >> TagAsUserFeatures() 
tag_item_id = ["item_id"] >> TagAsItemID()
tag_item_features = articles_features   >> TagAsItemFeatures() 
targets = ["Target"] >> AddMetadata(tags=[Tags.BINARY_CLASSIFICATION, "target"]) >> Dropna()


tag_to_select_unique_values = nvt.Workflow(tag_user_id + tag_user_features + tag_item_id + tag_item_features + targets)

train_dataset = Dataset(train_to_merlin)
test_dataset = Dataset(test_to_merlin)

train_df = tag_to_select_unique_values.fit_transform(train_dataset)
test_df = tag_to_select_unique_values.transform(test_dataset)

#train_transformed.isna().sum()
train_df.compute()

# Mettre en dataset 
train_df.to_parquet(DATA_FOLDER+'/Processed/train_numeric')
test_df.to_parquet(DATA_FOLDER+'/Processed/test_numeric')

df_user_id_row_processed = df_customers[['user_id', 'user_id_row']]
df_user_id_row_processed
df_user_id_row_processed.to_parquet(DATA_FOLDER+'/Processed/user_id_row_processed')
