==> 02_Training_Models_v4.ipynb <==
# Entraînement des modèles

- Ce notebook prépare les données pour l'inférence en enregistrant les features dand
- This notebook prepares the data for inference: features are saved in the feature store and item embeddings are stored in faiss.
- On entraine également le modèle de retrieval et de scoring pour l'inférence

import os
import gc
import glob
os.environ["TF_GPU_ALLOCATOR"]="cuda_malloc_async"

import nvtabular as nvt
from nvtabular.ops import *
import numpy as np

from merlin.models.utils.example_utils import workflow_fit_transform
from merlin.schema.tags import Tags
from merlin.models.utils.dataset import unique_rows_by_features

import merlin.models.tf as mm
from merlin.io.dataset import Dataset
import tensorflow as tf

# disable INFO and DEBUG logging everywhere
import logging
logging.disable(logging.WARNING)

INPUT_DATA_DIR = os.environ.get("INPUT_DATA_DIR", "/root/Data/Row/")
DATA_FOLDER = os.environ.get("DATA_FOLDER", "/root/Data/")
MODELS_FOLDER = os.environ.get("MODELS", "/root/Models/")
PROCESSED_FOLDER = os.environ.get("PROCESSED_FOLDER", "/root/Data/Processed/")
feature_repo_path = os.environ.get("FEAST_PATH", "/root/Data/feast_repo/feature_repo")

BATCH_SIZE = int(os.environ.get("BATCH_SIZE", 512))
from merlin.core.dispatch import get_lib
df_lib = get_lib()
df_lib

## Chargement des datasets issues du notebook 01

train_numeric = Dataset(os.path.join(PROCESSED_FOLDER, "train_numeric", "*.parquet"), part_size="128MB")
test_numeric = Dataset(os.path.join(PROCESSED_FOLDER, "test_numeric", "*.parquet"), part_size="128MB")
#dataset = Dataset(os.path.join(PROCESSED_FOLDER, "dataset_numeric", "*.parquet"))

train_numeric.compute().to_csv(os.path.join(DATA_FOLDER, "Plot/train_numeric"))
test_numeric.compute().to_csv(os.path.join(DATA_FOLDER, "Plot/test_numeric"))

## Export des features cliens et produits

### Préparation du Feast feature store

# for running this example on GPU, install the following libraries
#%pip install "feast==0.31" faiss-gpu

# for running this example on CPU, uncomment the following lines
# %pip install tensorflow-cpu "feast==0.31" faiss-cpu
# %pip uninstall cudf

#!rm -rf $DATA_FOLDER/feast_repo
!cd $DATA_FOLDER && feast init feast_repo

feature_repo_path = os.path.join(DATA_FOLDER, "feast_repo/feature_repo")
if os.path.exists(f"{feature_repo_path}/example_repo.py"):
    os.remove(f"{feature_repo_path}/example_repo.py")
if os.path.exists(f"{feature_repo_path}/data/driver_stats.parquet"):
    os.remove(f"{feature_repo_path}/data/driver_stats.parquet")

### Extraction des features

from datetime import datetime

#Extract unique rows for users
user_features = (
    unique_rows_by_features(train_numeric, Tags.USER, Tags.USER_ID)
    .compute()
    .reset_index(drop=True)
)

user_features["datetime"] = datetime.now()
user_features["datetime"] = user_features["datetime"].astype("datetime64[ns]")
user_features["created"] = datetime.now()
user_features["created"] = user_features["created"].astype("datetime64[ns]")

#Store in feast folder
user_features.to_parquet(
    os.path.join(feature_repo_path, "data", "user_features.parquet")
)


## Same for Items
item_features = (
    unique_rows_by_features(train_numeric, Tags.ITEM, Tags.ITEM_ID)
    .compute()
    .reset_index(drop=True)
)
item_features["datetime"] = datetime.now()
item_features["datetime"] = item_features["datetime"].astype("datetime64[ns]")
item_features["created"] = datetime.now()
item_features["created"] = item_features["created"].astype("datetime64[ns]")

item_features.to_parquet(
    os.path.join(feature_repo_path, "data", "item_features.parquet")
)

item_features.head()

user_features.head(2)

### Définition des features pour Feast

item_features.dtypes

file = open(os.path.join(feature_repo_path, "item_features.py"), "w")
file.write(
        """
from datetime import timedelta
from feast import Entity, Field, FeatureView, ValueType
from feast.types import Int32, Float32
from feast.infra.offline_stores.file_source import FileSource

item_features = FileSource(
    path="{}",
    timestamp_field="datetime",
    created_timestamp_column="created",
)

item = Entity(name="item_id", value_type=ValueType.INT32, join_keys=["item_id"],)

item_features_view = FeatureView(
    name="item_features",
    entities=[item],
    ttl=timedelta(0),
    schema=[
        Field(name="product_code", dtype=Int32),
        Field(name="prod_name", dtype=Int32),
        Field(name="product_type_no", dtype=Int32),
        Field(name="product_group_name", dtype=Int32),
        Field(name="graphical_appearance_no", dtype=Int32),
        Field(name="colour_group_code", dtype=Int32),
        Field(name="perceived_colour_value_id", dtype=Int32),
        Field(name="perceived_colour_master_id", dtype=Int32),
        Field(name="department_no", dtype=Int32),
        Field(name="index_code", dtype=Int32),
        Field(name="index_group_no", dtype=Int32),
    	Field(name="section_no", dtype=Int32),
    	Field(name="garment_group_no", dtype=Int32),
        Field(name="detail_desc", dtype=Int32),
        Field(name="count_30d_purchased", dtype=Int32),
        Field(name="count_7d_purchased", dtype=Int32),
        Field(name="Time_Weighted_Purchased", dtype=Float32),
    ],
    online=True,
    source=item_features,
    tags=dict(),
)
""".format(
            os.path.join(feature_repo_path, "data/", "item_features.parquet")
        )
    )
file.close()

user_features.dtypes

f = open(os.path.join(feature_repo_path, "user_features.py"), "w")
f.write(
    """
from datetime import timedelta
from feast import Entity, Field, FeatureView, ValueType
from feast.types import Int32, Float32
from feast.infra.offline_stores.file_source import FileSource

user_features = FileSource(
    path="{}",
    timestamp_field="datetime",
    created_timestamp_column="created",
)

user = Entity(name="user_id", value_type=ValueType.INT32, join_keys=["user_id"],)

user_features_view = FeatureView(
    name="user_features",
    entities=[user],
    ttl=timedelta(0),
    schema=[
        Field(name="FN", dtype=Int32),
        Field(name="Active", dtype=Int32),
        Field(name="club_member_status", dtype=Int32),
        Field(name="fashion_news_frequency", dtype=Int32),
        Field(name="age", dtype=Int32),
        Field(name="postal_code", dtype=Int32),
        Field(name="recency", dtype=Int32),
        Field(name="frequency", dtype=Float32),
        Field(name="amount", dtype=Float32),
        Field(name="popular_product_type", dtype=Int32),
        Field(name="2nd_popular_product_type", dtype=Int32),
        Field(name="popular_department_no", dtype=Int32),
        Field(name="2nd_popular_department_no", dtype=Int32),
        Field(name="popular_section_no", dtype=Int32),
        Field(name="2nd_popular_section_no", dtype=Int32),
        Field(name="last_product_code", dtype=Int32),
        Field(name="2nd_last_product_code", dtype=Int32),
        Field(name="last_product_type", dtype=Int32),
        Field(name="2nd_last_product_type", dtype=Int32),
    ],
    online=True,
    source=user_features,
    tags=dict(),
)
""".format(
        os.path.join(feature_repo_path, "data/", "user_features.parquet")
    )
)
f.close()

## Pre-processing avec Nvtabular

Définition des features et des workflows

#Item columns for Nvtabular workflow
cat_item_features = ['product_code', 'prod_name', 'product_type_no', 'graphical_appearance_no', 'colour_group_code', 'product_group_name',
       'perceived_colour_value_id', 'perceived_colour_master_id',
       'department_no', 'index_code', 'index_group_no', 'section_no',
       'garment_group_no', 'detail_desc', 'count_30d_purchased',
       'count_7d_purchased'] #'product_group_name'
float_item_features = ['Time_Weighted_Purchased']

#User columns
cat_user_features = ['FN', 'Active', 'club_member_status',
       'fashion_news_frequency', 'age', 'postal_code', 'popular_product_type', '2nd_popular_product_type',
       'popular_department_no', '2nd_popular_department_no',
       'popular_section_no', '2nd_popular_section_no', 'last_product_code',
       '2nd_last_product_code',
       'last_product_type', '2nd_last_product_type', 'recency']
float_user_features = ['frequency', 'amount']

#Item columns for Nvtabular workflow
cat_item_features = ['product_code', 'prod_name', 'product_type_no', 'graphical_appearance_no', 'colour_group_code', 'product_group_name',
       'perceived_colour_value_id', 'perceived_colour_master_id',
       'department_no', 'index_code', 'index_group_no', 'section_no',
       'garment_group_no', 'detail_desc']
float_item_features = ['Time_Weighted_Purchased', 'count_7d_purchased', 'count_30d_purchased']

#User columns
cat_user_features = ['FN', 'Active', 'club_member_status',
       'fashion_news_frequency', 'postal_code', 'popular_product_type', '2nd_popular_product_type',
       'popular_department_no', '2nd_popular_department_no',
       'popular_section_no', '2nd_popular_section_no', 'last_product_code',
       '2nd_last_product_code',
       'last_product_type', '2nd_last_product_type']
float_user_features = ['age', 'frequency', 'amount', 'recency']

#train_numeric.schema

%%time
from merlin.dag.ops.subgraph import Subgraph
category_temp_directory = os.path.join(DATA_FOLDER, "Processed/categories")


#Adding tags and categorify
user_id = ["user_id"] >> TagAsUserID() >> AddTags(tags='USER_ID')
user_features_cat = cat_user_features >> Categorify(dtype='int32', out_path=category_temp_directory) >> TagAsUserFeatures() >> AddTags(tags='CATEGORICAL')
user_features_float = float_user_features >> Normalize() >> ReduceDtypeSize() >> TagAsUserFeatures() >> AddTags(tags='CONTINUOUS')

item_id = ["item_id"] >> TagAsItemID() >> AddTags(tags='ITEM_ID')
item_features_cat = cat_item_features  >>  Categorify(dtype='int32', out_path=category_temp_directory) >> TagAsItemFeatures() >> AddTags(tags='CATEGORICAL')  
item_features_float = float_item_features >> Normalize() >> ReduceDtypeSize() >> TagAsItemFeatures() >> AddTags(tags='CONTINUOUS')

subgraph_item = Subgraph(
     "item", 
     Subgraph("item_id_sg", item_id) + 
    Subgraph("item_features_cat_sg", item_features_cat) + 
    Subgraph("item_features_float_sg", item_features_float))

subgraph_user = Subgraph(
     "user", 
     Subgraph("user_id_sg", user_id) + 
    Subgraph("user_features_cat_sg", user_features_cat) +
    Subgraph("user_features_float_sg", user_features_float))

targets = ["Target"] >> AddMetadata(tags=[Tags.BINARY_CLASSIFICATION, "target"])

outputs = subgraph_item + subgraph_user + targets
#outputs = outputs >> Dropna()

# Create and run workflow
workflow = nvt.Workflow(outputs)


workflow.fit_transform(train_numeric).to_parquet(DATA_FOLDER+'/Processed/train_processed')
workflow.transform(test_numeric).to_parquet(DATA_FOLDER+'/Processed/test_processed')

Enregistrement du workflow pour Triton plus tard

workflow.save(os.path.join(MODELS_FOLDER, "general_workflow"))

On peut supprimer les datasets dont on n'a plus besoin

import shutil
shutil.rmtree('/root/Data/Processed/dataset_numeric')
shutil.rmtree('/root/Data/Processed/train_numeric')
shutil.rmtree('/root/Data/Processed/test_numeric')

## Entraînement du modèle de Retrieval

train_processed = Dataset(os.path.join(PROCESSED_FOLDER, "train_processed", "*.parquet"))
test_processed = Dataset(os.path.join(PROCESSED_FOLDER, "test_processed", "*.parquet"))

#train_processed.schema

On ne sélectionne que les interactions positives car pour le two-tower model, le negative sampling se fait en batch

inputs = train_processed.schema.column_names
outputs = inputs >> Filter(f=lambda df: df["Target"] == 1)

nvt_wkflow = nvt.Workflow(outputs)
nvt_wkflow.fit(train_processed)

nvt_wkflow.transform(train_processed).to_parquet(
    output_path=os.path.join(PROCESSED_FOLDER, "train_tt_processed")
)

nvt_wkflow.transform(test_processed).to_parquet(
    output_path=os.path.join(PROCESSED_FOLDER, "test_tt_processed")
)

train_tt = Dataset(os.path.join(PROCESSED_FOLDER, "train_tt_processed", "*.parquet"), part_size="500MB")
test_tt = Dataset(os.path.join(PROCESSED_FOLDER, "test_tt_processed", "*.parquet"), part_size="500MB")

train_tt.schema.select_by_tag(Tags.ITEM_ID)

#train_tt.schema.excluding_by_tag(Tags.ITEM_ID)

from merlin.schema.schema import Domain, ColumnSchema, Schema

# Obtenez la colonne actuelle
current_item_id_col = train_tt.schema['item_id']

# Créez un nouveau domaine
new_domain = Domain(min=0, max=999999)  # Ajustez ces valeurs selon vos besoins

# Créez une nouvelle instance de ColumnSchema avec le nouveau domaine
new_item_id_col = ColumnSchema(
    name="item_id",
    dtype=current_item_id_col.dtype,
    tags=current_item_id_col.tags + [Tags.ITEM_ID],  # Assurez-vous que Tags.ITEM_ID est présent
    properties={
        **current_item_id_col.properties,
        "domain": new_domain.__dict__,
        "is_categorical": True
    },
    is_list=current_item_id_col.is_list,
    is_ragged=current_item_id_col.is_ragged
)


new_item_id_col

schema_without_id = train_tt.schema.excluding_by_tag(Tags.ITEM_ID)
# Créez un nouveau schéma en supprimant l'ancienne colonne 'item_id' et en ajoutant la nouvelle
updated_schema = (schema_without_id + Schema([new_item_id_col])  # Ajoute la nouvelle colonne 'item_id'
)
updated_schema

#schema = train_tt.schema.select_by_tag([Tags.ITEM_ID, Tags.USER_ID, Tags.ITEM, Tags.USER])
schema =train_tt.schema.select_by_tag([Tags.ITEM_ID, Tags.USER_ID, Tags.ITEM, Tags.USER])
train_tt.schema = schema
test_tt.schema = schema

#schema

tower_dim = 64 

# create user schema using USER tag
user_schema = schema.select_by_tag(Tags.USER)
# create user (query) tower input block
user_inputs = mm.InputBlockV2(user_schema)

# create user (query) encoder block
query = mm.Encoder(user_inputs, mm.MLPBlock([128, tower_dim], no_activation_last_layer=True))

# create item schema using ITEM tag
item_schema = schema.select_by_tag(Tags.ITEM)
# create item (candidate) tower input block
item_inputs = mm.InputBlockV2(item_schema)
# create item (candidate) encoder block
candidate = mm.Encoder(item_inputs, mm.MLPBlock([128, tower_dim], no_activation_last_layer=True))

model_tt = mm.TwoTowerModelV2(query, candidate)
model_tt.compile(optimizer="adam", run_eagerly=False, loss="categorical_crossentropy", metrics=[mm.RecallAt(10), mm.NDCGAt(10)])

4096//2

%%time

model_tt.fit(train_tt, validation_data=test_tt, batch_size=4096//4, epochs=5)

import shutil
shutil.rmtree('/root/Data/Processed/test_tt_processed')
shutil.rmtree('/root/Data/Processed/train_tt_processed')

Sauvegarde du modèle de retrieval

query_tower = model_tt.query_encoder
query_tower.save(os.path.join(MODELS_FOLDER, "query_tower"))

### Sauvegarde des embeddings produits

item_features = Dataset(os.path.join(feature_repo_path, "data/item_features.parquet"))

item_features.schema

List_item_features = ['item_id', 'product_code', 'prod_name', 'product_type_no',
       'product_group_name', 'graphical_appearance_no', 'colour_group_code',
       'perceived_colour_value_id', 'perceived_colour_master_id',
       'department_no', 'index_code', 'index_group_no', 'section_no',
       'garment_group_no', 'detail_desc', 'count_30d_purchased',
       'count_7d_purchased', 'Time_Weighted_Purchased']
print(len(List_item_features))

On fait passer tous les items dans la branche item du two-tower model pour pré-calculer les embeddings

from nvtabular import Workflow

#nvt_workflow = Workflow.load(os.path.join(MODELS_FOLDER, 'general_workflow'))
item_subgraph = workflow.get_subworkflow("item")
#user_features = user_attributes >> TransformWorkflow(user_subgraph)

from merlin.systems.dag.ops.tensorflow import PredictTensorflow
from merlin.systems.dag.ops.workflow import TransformWorkflow

workflow_item_embedding =  nvt.Workflow(["item_id"] + (List_item_features >> TransformWorkflow(item_subgraph) >> PredictTensorflow(model_tt.candidate_encoder)))
item_embeddings = workflow_item_embedding.fit_transform(item_features).to_ddf().compute()
item_embeddings.tail()

Sauvegarde des embeddings

# save to disk
item_embeddings.to_parquet(os.path.join(PROCESSED_FOLDER, "item_embeddings.parquet"))

## Entraînement du modèle de scoring

# define schema object
schema_dlrm = train_processed.schema

target_column = schema_dlrm.select_by_tag(Tags.TARGET).column_names[0]
target_column

batch_size = 16 * 1024
LR = 0.03
metrics = [tf.keras.metrics.AUC(name="auc"),tf.keras.metrics.Precision(name='precision'), tf.keras.metrics.Recall(name='recall')] #mm.RecallAt(10), mm.NDCGAt(10), mm.PrecisionAt(10)]

model = mm.DLRMModel(
    schema_dlrm,
    embedding_dim=64,
    bottom_block=mm.MLPBlock([128, 64]),
    top_block=mm.MLPBlock([128, 64, 32]),
    prediction_tasks=mm.BinaryOutput('Target'),
)


%%time
opt = tf.keras.optimizers.legacy.Adagrad(learning_rate=LR*4)
model.compile(optimizer=opt, run_eagerly=False, metrics=metrics)
model.fit(train_processed, validation_data=test_processed, batch_size=batch_size, epochs=8)

from merlin.models.utils.example_utils import workflow_fit_transform, save_results

results_path = os.path.join( "results_models/results.txt")
if os.path.isfile(results_path):
    os.remove(results_path)

def save_results_custom(model_name, model, out_path):
    """a funct to save validation accuracy results in a text file"""
    with open(out_path, "a") as f:
        f.write(model_name)
        f.write("\n")
        for key, value in model.history.history.items():
            if "val_auc" in key:
                f.write("%s:%s\n" % (key, value[0]))
            if "val_loss" in key:
                f.write("%s:%s\n" % (key, value[0]))
            if "val_precision" in key:
                f.write("%s:%s\n" % (key, value[0]))
            if "val_recall" in key:
                f.write("%s:%s\n" % (key, value[0]))


save_results_custom("DLRM", model, results_path)

Enregistrement du modèle

model.save(os.path.join(MODELS_FOLDER, "dlrm"))

## Test NCF

schema = schema_dlrm

Test Wide & Deep

cat_schema = schema.select_by_tag(Tags.CATEGORICAL)

one_hot_encoding = mm.CategoryEncoding(cat_schema, sparse=True, output_mode="one_hot") # One-hot encoding

features_crossing = mm.HashedCrossAll(
        cat_schema,
        num_bins=100, # The crossed features will be hashed to this number of bins
        max_level=2,
        output_mode="one_hot",
        sparse=True,
        ignore_combinations=[["item_id", "item_category"], 
                             ["item_id", "item_brand"]]
    )

wide_preprocessing_blocks = mm.ParallelBlock([
                                              one_hot_encoding, 
                                              features_crossing
                                             ],
                                             aggregation="concat")

deep_part = mm.MLPBlock([128, 64, 32])

model = mm.WideAndDeepModel(
        schema,
        wide_schema=cat_schema,
        deep_schema=schema,
        wide_preprocess=wide_preprocessing_blocks,
        deep_block=deep_part,
        prediction_tasks=mm.BinaryOutput(target_column),
    )

%%time
opt = tf.keras.optimizers.legacy.Adagrad(learning_rate=LR*5)
model.compile(optimizer=opt, run_eagerly=False, metrics=metrics)
model.fit(train_processed, validation_data=test_processed, batch_size=batch_size//2, epochs=8)

val_loss: 0.5028 - val_auc: 0.8365 - val_precision: 0.7053 - val_recall: 0.8006 - val_regularization_loss: 0.0000e+00 - val_loss_batch: 0.4982

#save_results_custom("Wide&Deep", model, results_path)

#model.save(os.path.join(MODELS_FOLDER, "wandd"))

model = mm.DCNModel(
    schema,
    depth=1,
    deep_block=mm.MLPBlock([64, 32]),
    prediction_tasks=mm.BinaryOutput(target_column),
)

%%time

opt = tf.keras.optimizers.legacy.Adagrad(learning_rate=LR*10)
model.compile(optimizer=opt, run_eagerly=False, metrics=metrics)
model.fit(train_processed, validation_data=test_processed, batch_size=batch_size//2, epochs=8)

save_results_custom("DCN", model, results_path)

model.save(os.path.join(MODELS_FOLDER, "DCN"))

# uses default embedding_dim = 64
model = mm.Model.from_block(mm.MLPBlock([64, 32]),
    schema, prediction_tasks=mm.BinaryOutput(target_column)
)


%%time

opt = tf.keras.optimizers.legacy.Adagrad(learning_rate=LR*4)
model.compile(optimizer=opt, run_eagerly=False, metrics=metrics)
model.fit(train_processed, validation_data=test_processed, batch_size=batch_size, epochs=10)
