==> 03_Deployment_v4.ipynb <==
# Deploiement de la pipeline de recommandation personalisée

Ce notebook a pour but de créer un ensemble triton pour déployer le système de recommandation

#%pip install "feast<0.31" faiss-gpu
#!pip install seedir

import os
import numpy as np
import pandas as pd
import feast
import seedir as sd
from nvtabular import ColumnSchema, Schema

from merlin.systems.dag.ensemble import Ensemble
from merlin.systems.dag.ops.softmax_sampling import SoftmaxSampling
from merlin.systems.dag.ops.tensorflow import PredictTensorflow
from merlin.systems.dag.ops.unroll_features import UnrollFeatures
from merlin.systems.triton.utils import send_triton_request
from merlin.systems.dag.ops.workflow import TransformWorkflow

## Enregistrement des features dans le feature store

INPUT_DATA_DIR = os.environ.get("INPUT_DATA_DIR", "/root/Data/Row/")
DATA_FOLDER = os.environ.get("DATA_FOLDER", "/root/Data/")
MODELS_FOLDER = os.environ.get("MODELS", "/root/Models/")
PROCESSED_FOLDER = os.environ.get("PROCESSED_FOLDER", "/root/Data/Processed/")
feature_repo_path = os.environ.get("FEAST_PATH", "/root/Data/feast_repo/feature_repo")

import seedir as sd

feature_repo_path = os.path.join(feature_repo_path)
sd.seedir(
    feature_repo_path,
    style="lines",
    itemlimit=10,
    depthlimit=3,
    #exclude_folders=".ipynb_checkpoints",
    sort=True,
)

Enregistrementdes données dans le feature store.
Le Feast feature registry est le catalogue centrale des définitions des features et des metadata.

Chargement des features depuis l'offline store à l'online store

%cd $feature_repo_path
!find . -name ".ipynb_checkpoints" -exec rm -r {} +
!feast apply

#!find . -name ".ipynb_checkpoints" -exec rm -r {} +

!feast materialize 1995-01-01T01:01:01 2025-01-01T01:01:01

# set up the base dir to for feature store
sd.seedir(os.path.join(feature_repo_path), style='lines', itemlimit=10, depthlimit=5, sort=True) #exclude_folders=['.ipynb_checkpoints', '__pycache__']

Création du client du feast

feature_store = feast.FeatureStore(feature_repo_path)

Definition des chemins

if not os.path.isdir(os.path.join(DATA_FOLDER, 'faiss_index')):
    os.makedirs(os.path.join(DATA_FOLDER, 'faiss_index'))

faiss_index_path = os.path.join(DATA_FOLDER, 'faiss_index', "index.faiss")
retrieval_model_path = os.path.join(MODELS_FOLDER, "query_tower/")
ranking_model_path = os.path.join(MODELS_FOLDER, "dlrm/")

Create de l'index Faiss à partir des embeddings

from merlin.systems.dag.ops.faiss import QueryFaiss, setup_faiss 

item_embeddings = pd.read_parquet(os.path.join(DATA_FOLDER, "Processed/item_embeddings.parquet"))
setup_faiss(item_embeddings, faiss_index_path, embedding_column="output_1")

## Definition de la pipeline de recommandation complète avec Nvtabular

import warnings
import logging
logging.basicConfig(level=logging.DEBUG, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')

class EnhancedTimedQueryFeast(QueryFeast):
    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)
        self.log_file = open('queryfeast_logs.txt', 'w')

    def log_message(self, message):
        print(message, file=self.log_file, flush=True)
        print(message, file=sys.stdout, flush=True)

    def transform(self, col_selector: List[str], df: "cudf.DataFrame") -> "cudf.DataFrame":
        start_time = time.time()
        self.log_message(f"Début de QueryFeast.transform() à {start_time}")
        self.logger.error(f"Debut Feast {time.time()}")
        result = super().transform(col_selector, df)
        
        end_time = time.time()
        duration = end_time - start_time
        self.log_message(f"Fin de QueryFeast.transform(). Durée: {duration:.4f} secondes")
        self.logger.error(f"Fin Feast {time.time()}")
        return result

    def compute_output_schema(self, input_schema: Schema, col_selector: List[str], prev_output_schema: Schema = None) -> Schema:
        start_time = time.time()
        self.log_message(f"Début de QueryFeast.compute_output_schema() à {start_time}")
        
        result = super().compute_output_schema(input_schema, col_selector, prev_output_schema)
        
        end_time = time.time()
        duration = end_time - start_time
        self.log_message(f"Fin de QueryFeast.compute_output_schema(). Durée: {duration:.4f} secondes")
        
        return result

    def __del__(self):
        if hasattr(self, 'log_file'):
            self.log_file.close()


from merlin.core.dispatch import make_df
from merlin.systems.dag.ops.feast import QueryFeast
from nvtabular import ColumnSelector, Workflow, Dataset


#Test input
request = make_df({"user_id": [11]})
request["user_id"] = request["user_id"].astype(np.int32)
test_dataset = Dataset(request)


# Embedding retrieval >> TimerOperator('Début du worklow')
user_attributes = ["user_id"] >> EnhancedTimedQueryFeast.from_feature_view(
    store=feature_store,
    view="user_features",
    column="user_id",
    include_id=True,
)

# >> TimerOperator('Récupération des attributes') 
nvt_workflow = Workflow.load(os.path.join(MODELS_FOLDER, 'general_workflow'))
user_subgraph = nvt_workflow.get_subworkflow("user")
user_features = user_attributes >> TransformWorkflow(user_subgraph) 

configure_tensorflow()

topk_retrieval = int(
    os.environ.get("topk_retrieval", "100")
)
# Predictensorflow 1 
retrieval = (
    user_features
    #>> TimerOperator('Transformation des features user') 
    >> PredictTensorflow(retrieval_model_path)
    >> QueryFaiss(faiss_index_path, topk=topk_retrieval)
)

# >> TimerOperator('Retrieval Two Tower') 
item_attributes = retrieval["candidate_ids"] >>  QueryFeast.from_feature_view(
    store=feature_store,
    view="item_features",
    column="candidate_ids",
    #output_prefix="item",
    include_id=True,
)


output = TEST_workflow.fit_transform(test_dataset)


print(output.to_ddf().compute())

Premier opérateur : récupération des features brut d'un client à partir d'un user_id

from merlin.systems.dag.ops.feast import QueryFeast 

user_attributes = ["user_id"] >> EnhancedTimedQueryFeast.from_feature_view(
    store=feature_store,
    view="user_features",
    column="user_id",
    include_id=True,
)

Préparation des données pour le Two Tower model

from nvtabular import Workflow
# Premier workflow
nvt_workflow = Workflow.load(os.path.join(MODELS_FOLDER, 'general_workflow'))
user_subgraph = nvt_workflow.get_subworkflow("user")
user_features = user_attributes  >>  TransformWorkflow(user_subgraph)

# prevent TF to claim all GPU memory
from merlin.dataloader.tf_utils import configure_tensorflow

configure_tensorflow()

On récupère les items candidats en comparant l'embedding du client aux embeddings des produits via faiss.

On récupère ensuite les features de ces items dans feast.

topk_retrieval = int(
    os.environ.get("topk_retrieval", "100")
)
# Predictensorflow 1 
retrieval = (
    user_features
    >> PredictTensorflow(retrieval_model_path)
    >> QueryFaiss(faiss_index_path, topk=topk_retrieval)
)

item_attributes = retrieval["candidate_ids"]   >>  EnhancedTimedQueryFeast.from_feature_view(
    store=feature_store,
    view="item_features",
    column="candidate_ids",
    #output_prefix="item",
    include_id=True,
)

item_subgraph = nvt_workflow.get_subworkflow("item")
item_features = item_attributes  >> TransformWorkflow(item_subgraph)

user_features_to_unroll = [
    "user_id",
    "FN",
    "Active",
    "club_member_status",
    "fashion_news_frequency",
    "age",
    "postal_code",
    "recency",
    "frequency",
    "amount", 'popular_product_type', '2nd_popular_product_type', 'popular_department_no', '2nd_popular_department_no', 'popular_section_no', '2nd_popular_section_no', 'last_product_code', '2nd_last_product_code', 'last_product_type', '2nd_last_product_type']

combined_features = item_features  >> UnrollFeatures(
    "item_id", user_features[user_features_to_unroll]
)

Scoring avec le modèle DLRM sauvegardé précédemment

ranking = combined_features >> PredictTensorflow(ranking_model_path)

top_k=12
ordering = combined_features["item_id"] >>  SoftmaxSampling(
    relevance_col=ranking["Target/binary_output"], topk=top_k, temperature=0.00000001
) 

from merlin.core.dispatch import make_df
from merlin.systems.dag.ops.feast import QueryFeast
from nvtabular import ColumnSelector, Workflow, Dataset

#Test input
request = make_df({"user_id": [11]})
request["user_id"] = request["user_id"].astype(np.int32)
test_dataset = Dataset(request)

# Exécuter la pipeline sur les données de test
similarity_workflow = Workflow(ordering)
output = similarity_workflow.transform(test_dataset)


print(output.to_ddf().compute())

## Sauvegarde de la pipeline complète

Create folder to export graphs and config

if not os.path.isdir("/root/Triton_models"):
    os.makedirs(os.path.join('/root/Triton_models'))

Création du schema de l'input qu'on va donner à Triton.

request_schema = Schema(
    [
        ColumnSchema("user_id", dtype=np.int32),
    ]
)
request_schema

Création de l'ensemble python (la pipeline au complet)

%%time
# define the path where all the models and config files exported to
export_path = os.path.join('/root/Triton_models_test_operateur_perso')

ensemble = Ensemble(ordering, request_schema)
ens_config, node_configs = ensemble.export(export_path)

# return the output column name
outputs = ensemble.graph.output_schema.column_names
print(outputs)

sd.seedir( os.path.join('/root/Triton_models_test_operateur_perso'), style='lines', itemlimit=10, depthlimit=5, sort=True) #exclude_folders=['.ipynb_checkpoints', '__pycache__']

## Démarrage du Triton Server

Let's clean some useless files

import shutil
export_path = os.path.join('/root/Triton_models_test_operateur_perso')

def remove_checkpoints(dir_path):
    for root, dirs, files in os.walk(dir_path):
        for dir_name in dirs:
            if dir_name == '.ipynb_checkpoints':
                dir_to_remove = os.path.join(root, dir_name)
                print(f"Removing: {dir_to_remove}")
                shutil.rmtree(dir_to_remove)

remove_checkpoints(export_path)

#sd.seedir(export_path, style='lines', itemlimit=10, depthlimit=5, sort=True)

Pour démarrer Triton Inference Server, il faut executer la commande suivande dans un terminale dans ce conteneur :

tritonserver --model-repository=/root/Triton_models/ --backend-config=tensorflow,version=2

Il faut attendre d'avoir tous les modèles ready dans le terminal avant d'executer les test (cela peut être vraiment long 20-30 min)

# read in data for request

from merlin.core.dispatch import make_df

# create a request to be sent to TIS
request = make_df({"user_id": [9]})
request["user_id"] = request["user_id"].astype(np.int32)
print(request)

outputs = ['ordered_ids', 'ordered_scores']

request_schema = Schema(
    [
        ColumnSchema("user_id", dtype=np.int32),
    ]
)

response = send_triton_request(request_schema, request, outputs)
response

from merlin.core.dispatch import make_df

# create a request to be sent to TIS
request = make_df({"user_id": [9]})
request["user_id"] = request["user_id"].astype(np.int32)
print(request)

outputs = ['ordered_ids', 'ordered_scores']

request_schema = Schema(
    [
        ColumnSchema("user_id", dtype=np.int32),
    ]
)

%%time
response = send_triton_request(request_schema, request, outputs)
response
